+++ src/stories/TabAccountSetting.stories.tsx
@@ -1,9 +1,9 @@
 import { Meta, StoryObj } from "@storybook/react";
-import TabAccountSetting from "@/components/TabAccountSetting";
+import TabAccountSetting from "@/layout/AccountSettingsLayout";
 
 const meta: Meta<typeof TabAccountSetting> = {
-  title: "components/TabAccountSetting",
-  component: TabAccountSetting,
+	title: "components/TabAccountSetting",
+	component: TabAccountSetting,
 };
 
 export default meta;
